Refactor directory structure and code
