+++ src/Exception/Client/ConnectionException.php
@@ -4,12 +4,12 @@
  * 
  * Exception thrown when WebSocket connection fails.
  * 
- * @package     Sockeon\Sockeon\Client\Exceptions
- * @author      Xentixar
+ * @package     Sockeon\Sockeon
+ * @author      Sockeon
  * @copyright   Copyright (c) 2025
  */
 
-namespace Sockeon\Sockeon\Client\Exceptions;
+namespace Sockeon\Sockeon\Exception\Client;
 
 class ConnectionException extends WebSocketException
 {
