+++ src/Exception/Client/HandshakeException.php
@@ -4,12 +4,12 @@
  * 
  * Exception thrown when WebSocket handshake fails.
  * 
- * @package     Sockeon\Sockeon\Client\Exceptions
- * @author      Xentixar
+ * @package     Sockeon\Sockeon
+ * @author      Sockeon
  * @copyright   Copyright (c) 2025
  */
 
-namespace Sockeon\Sockeon\Client\Exceptions;
+namespace Sockeon\Sockeon\Exception\Client;
 
 class HandshakeException extends WebSocketException
 {
