+++ src/Exception/Client/WebSocketException.php
@@ -4,12 +4,12 @@
  * 
  * Base exception class for WebSocket client errors.
  * 
- * @package     Sockeon\Sockeon\Client\Exceptions
- * @author      Xentixar
+ * @package     Sockeon\Sockeon
+ * @author      Sockeon
  * @copyright   Copyright (c) 2025
  */
 
-namespace Sockeon\Sockeon\Client\Exceptions;
+namespace Sockeon\Sockeon\Exception\Client;
 
 use Exception;
